// apps/web/src/generated/types/projects/GetProjectsByProjectIdResearchers.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type GetProjectsByProjectIdResearchersPathParams = {
    /**
     * @type string
    */
    projectId: string;
};

/**
 * @description Response for status 200
*/
export type GetProjectsByProjectIdResearchers200 = {
    /**
     * @type string
    */
    researcherId: string;
    /**
     * @type string
    */
    name: string;
    /**
     * @type string
    */
    email: string;
    /**
     * @type string
    */
    globalRole: string;
    /**
     * @type string
    */
    projectRole: string;
}[];

export type GetProjectsByProjectIdResearchersQueryResponse = GetProjectsByProjectIdResearchers200;

export type GetProjectsByProjectIdResearchersQuery = {
    Response: GetProjectsByProjectIdResearchers200;
    PathParams: GetProjectsByProjectIdResearchersPathParams;
    Errors: any;
};